+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - [#50](https://github.com/n15g/badger/pull/50) - Character import and export, and log parser.
+- [#53](https://github.com/n15g/badger/pull/53) - Collect multiple badges simultaneously from the badge list.
 
 ### Changed